Mv OpPayloadTypes to reth-optimism-payload-builder
