+++ crates/optimism/payload/Cargo.toml
@@ -50,3 +50,4 @@ derive_more.workspace = true
 tracing.workspace = true
 thiserror.workspace = true
 sha2.workspace = true
+serde.workspace = true
