+++ crates/optimism/payload/src/lib.rs
@@ -15,10 +15,37 @@ pub mod builder;
 pub use builder::OpPayloadBuilder;
 pub mod error;
 pub mod payload;
+use op_alloy_rpc_types_engine::OpExecutionData;
 pub use payload::{OpBuiltPayload, OpPayloadAttributes, OpPayloadBuilderAttributes};
 mod traits;
+use reth_optimism_primitives::{OpBlock, OpPrimitives};
+use reth_payload_primitives::{BuiltPayload, PayloadTypes};
+use reth_primitives_traits::{NodePrimitives, SealedBlock};
 pub use traits::*;
 pub mod validator;
 pub use validator::OpExecutionPayloadValidator;
 
 pub mod config;
+
+/// A default payload type for [`OpEngineTypes`]
+#[derive(Debug, Default, Clone, serde::Deserialize, serde::Serialize)]
+#[non_exhaustive]
+pub struct OpPayloadTypes<N: NodePrimitives = OpPrimitives>(core::marker::PhantomData<N>);
+
+impl<N: NodePrimitives> PayloadTypes for OpPayloadTypes<N>
+where
+    OpBuiltPayload<N>: BuiltPayload<Primitives: NodePrimitives<Block = OpBlock>>,
+{
+    type ExecutionData = OpExecutionData;
+    type BuiltPayload = OpBuiltPayload<N>;
+    type PayloadAttributes = OpPayloadAttributes;
+    type PayloadBuilderAttributes = OpPayloadBuilderAttributes<N::SignedTx>;
+
+    fn block_to_payload(
+        block: SealedBlock<
+            <<Self::BuiltPayload as BuiltPayload>::Primitives as NodePrimitives>::Block,
+        >,
+    ) -> Self::ExecutionData {
+        OpExecutionData::from_block_unchecked(block.hash(), &block.into_block())
+    }
+}
